Merge all must-haves and should-haves into main

See merge request cse2115/2021-2022/sem-group-08b/sem-repo-08b!32

+++ bookings_microservice/src/main/java/nl/tudelft/sem/template/services/BookingService.java
@@ -38,12 +38,33 @@ public class BookingService {
     }
 
     public void updateBooking(Long id, Booking booking) {
-        bookingRepository.deleteById(id);
         bookingRepository.save(booking);
     }
 
-    public void deleteBooking(Long id) {
-        bookingRepository.deleteById(id);
+    /**
+     * Method that allows a user to edit their own bookings.
+     *
+     * @param userId the id of the user
+     * @param id the id of the booking
+     * @param booking the booking to be edited
+     */
+    public void updateMyBooking(String userId, Long id, Booking booking) {
+        if (booking.getBookingOwner().equals(userId)) {
+            bookingRepository.save(booking);
+        }
+    }
+
+    /**
+     * Method that allows a user to delete their own bookings.
+     *
+     * @param userId the id of the user
+     * @param id the id of the booking
+     */
+    public void deleteMyBooking(String userId, Long id) {
+        Booking booking = getBooking(id);
+        if (booking.getBookingOwner().equals(userId)) {
+            bookingRepository.deleteById(id);
+        }
     }
 
     /** get the schedule for the user using the specified sorting strategy.
